V 4.0/requirements.txt
bidict==0.23.1
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.7.0
eventlet==0.37.0
Flask==3.0.3
Flask-SocketIO==5.4.1
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-speech==2.27.0
google-cloud-texttospeech==2.17.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pillow==10.2.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydub==0.25.1
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.4
requests==2.31.0
rsa==4.9
simple-websocket==1.0.0
urllib3==2.2.3
Werkzeug==3.0.4
wsproto==1.2.0
